requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
django-ninja==1.3.0
django-crontab
django-redis==5.4.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
mypy-extensions==1.0.0
openai==1.59.4
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
redis==5.0.8
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
whitenoise==6.8.2
pytest-django==4.10.0
pytest-asyncio==0.26.0
django-extensions==3.2.3